Resolve issues with serializing NaN values
Previously, serializing NaN values (0/0) would cause an error as the
check for fractional values wouldn't consider NaNs as needing to be
encoded as floating point values, causing them to instead be written out
as single-byte integers and causing errors in later string.char calls.

This change improves the handling of NaNs and non-finite values so that,
at the very least, they don't error in the serialization/deserialization
process and are now actively tested.

The following cases now correctly serialize and deserialize without
errors:

  - Infinity
  - -Infinity
  - NaN

The following cases have improved such that they won't error upon
serialization, but won't round trip to their original values:

  - QNaN (will deserialize as NaN)

A few quirks have been observed however due to the WoW environment
being compiled with floating point optimizations (/fp:fast) compared to
a typical Lua 5.1 interpreter that isn't.

The first quirk is that NaN values are inequal to themselves outside the
game, but inside the WoW environment all comparisons involving NaNs are
actually true.

The next quirk is that the stringified forms of various non-finite values
differ between the two environments, which would have led to a few subtle
serialization differences - one example is that the string representation
of infinity in WoW is "inf" and can be converted to/from a string via
tonumber, whereas elsewhere it'll be "#1.INF" and *can't* roundtrip
through tostring/tonumber.

In terms of what's not implemented, the main one is support for quiet
NaNs (QNaNs) as generated by the expression "-(0/0)". Handling of these
is improved in the sense that they no longer error, but they won't round
trip though the library and will deserialize as standard NaNs. These
should be rare enough to not matter for now, and having them not error
is probably good enough.

Signalling NaNs (SNaNs) are untested as at present I'm unsure if it's
possible to get a representation of them in Lua whatsoever.

Finally, serializing tables where the key is a NaN is left unsupported
entirely. Outside of the WoW environment such a key is actually invalid,
whereas ingame it works but actually seems to hard crash the client upon
serialization.