Improve serialization of non-finite values (#3)
Serializing non-finite values such as NaNs and infinity would cause an error as the check for fractional values in the `number` type writer wouldn't consider such inputs, causing them to instead be written out as single-byte integers and causing errors in later string.char calls. The following test cases would manifest the error shown below:

```lua
LibSerialize:Serialize(0/0) -- This one would work outside of the WoW environment, but not inside.
LibSerialize:Serialize(math.huge)
LibSerialize:Serialize(-math.huge)
```

```
1x integer overflow attempting to store -nan(ind)
[string "=[C]"]: ?
[string "@LibSerialize\LibSerialize-1.lua"]:541: in function <LibSerialize\LibSerialize.lua:539>
[string "@LibSerialize\LibSerialize-1.lua"]:902: in function `_WriteInt'
[string "@LibSerialize\LibSerialize-1.lua"]:898: in function `_WriteByte'
[string "@LibSerialize\LibSerialize-1.lua"]:974: in function `writeFn'
[string "@LibSerialize\LibSerialize-1.lua"]:893: in function `_WriteObject'
[string "@LibSerialize\LibSerialize-1.lua"]:1248: in function <LibSerialize\LibSerialize.lua:1229>
```

This change improves the handling of non-finite values so that, at the very least, they don't error in the serialization/deserialization process and are now actively tested.

The following cases now correctly serialize and deserialize to their original values without errors:

  - Infinity
  - -Infinity
  - NaN

The following cases have improved such that they won't error upon serialization, but won't round trip to their original values:

  - QNaN (will deserialize as NaN)

A few quirks have been observed due to the WoW environment being compiled with floating point optimizations (`/fp:fast`) compared to a typical Lua 5.1 interpreter that isn't.

The first quirk is that NaN values are inequal to themselves outside the game, but inside the WoW environment all comparisons involving NaNs are actually true.

The next quirk is that the stringified forms of various non-finite values differ between the two environments, which would have led to a few subtle serialization differences - one example is that the string representation of infinity in WoW is `"inf"` and can be converted back from a string via tonumber, whereas elsewhere it'll be `"#1.INF"` and *can't* roundtrip through tostring/tonumber.

In terms of what's not implemented, the main one is support for quiet NaNs (qNaNs) as generated by the expression `-(0/0)`. Handling of these is improved in the sense that they no longer error, but they won't round trip though the library and will deserialize as standard NaNs. These should be rare enough to not matter for now, and having them not error is probably good enough.

Signalling NaNs (sNaNs) are untested as at present I'm unsure if it's possible to get a representation of them in Lua whatsoever.

Finally, serializing tables where the key is a NaN is left unsupported entirely. Outside of the WoW environment such a key is actually invalid and will error upon insertion, whereas ingame the insertion works but serialization seems to hard crash the client. I feel supporting this particular case is _probably_ out of scope given the stock VM has an explicit error condition in place and it's not likely Blizzard actively intended for this to be supported ingame.