Swap to Reader protocol implementation
This removes the added "options" parameter for the deserialization
APIs in favor of instead allowing users to just supply a custom
Reader object via the "input" directly.

The Reader interface matches - minus case differences - that which
was exposed in previous commits; should the given "input" be an
indexable value which has assignments on the "ReadBytes" and "AtEnd"
keys those will replace the default operations performed internally
when processing the stream.

Backwards compatibility was not harmed in the making of this commit,
and the test cases have been updated and cleaned up appropriately -
so should serve quite well as examples too now.