Disable failing compat test for NaN values
The serialization behaviour of NaN values in certain environments
changed between releases a bit.

Prior to v1.1.2 NaNs weren't correctly detected in the in-game
environment, therefore any test case that compares the serialization
of NaNs in a version older than that to a version newer will always
fail in such an environment.

Therefore - the NaN test is now only compatibility tested between v1.1.2
and upwards.