Merge async APIs and generic reader/writer support (#11)
* Add async serialize option via coroutines (#4)

* Add chunks mode for coroutine processing

* Requested updates

* Split sync/async functions, others

* Further updates

* Implement support for custom readers and writers (#12)

* Implement generic support for custom readers

This permits a new options table to be supplied to the Deserialize
and DeserializeValue functions on the library which may contain two
keys that control how the library reads bytes from the stream for
deserialization, and how it detects the end-of-stream state.

As-is this implementation is fully backwards compatible and allows
users of the library to throw processing into a coroutine for large
inputs if required, meeting some use cases that have prompted similar
PRs whilst also being more flexible and much simpler.

The new signatures for Deserialize and DeserializeValue are as follows:

  function LibSerialize:DeserializeValue(input[, options])
  function LibSerialize:Deserialize(input[, options])

The options parameter, if supplied, is expected to be a table that
may contain the following optional keys:

  - readBytes
  - atEnd

The "readBytes" key if present is expected to point to a function
of the signature "readBytes(input, i, j)" and must return a string
whose length is equal to (j - i) + 1). The "input" parameter is the
same as-supplied to Deserialize/DeserializeValue verbatim. The "i"
and "j" parameter represent an inclusive range of indices ([i, j])
typically as byte offsets within the input value. This convention is
identical to that provided by many Lua functions, such as string.sub,
table.concat, and unpack.

The "atEnd" key if present is expected to point to a function of the
signature "atEnd(input, i)". As with "readBytes", the "input" parameter
is whatever value was supplied to Deserialize/DeserializeValue, and the
"i" parameter represents the offset of the stream.

By default if neither of these keys are supplied, or no options table
is given, then the implementation will just default to using string.sub
for "readBytes" and for "atEnd" will test if the offset exceeds the
length of the input as-provided by the length operator.

Examples are provided in the test cases, where we test the following
scenarios:

  - A case where "input" may have additional data after the payload that
    we don't want deserializing.
  - A case where the "input" is a table of character bytes.
  - A case where we yield from within "readBytes" and feed data from the
    main thread, simulating receiving data with a potentially unknown
    bound.

* Implement support for custom writers

This mirrors recent changes submitted for the deserialization API and
provides a generic system for custom writer support to control how and
where the byte stream generated during serialization is written. The
change as-is is fully backwards compatible and passes the existing and
expanded test cases.

Primary use cases would align with existing async serialization attempts
but don't put the complexity of thread management or throttling logic
into the library, instead leaving the minutia of how those aspects are
handled to callers. It also enables some other advanced use cases such
as streaming the data to a destination without buffering it fully
beforehand, which could enable efficiencies elsewhere.

* Convert generic reader/writer to use objects instead of closures

* Hoist async into reader/writer layer instead of serialize/deserialize

* Rename function to GenericReader_AtEnd

* Expand tests, add example for reader/writer protocol

* Update readme

* Fix formatting

* Globals cleanup

---------

Co-authored-by: Mark Bosley <mark.bosley@gmail.com>
Co-authored-by: Daniel Yates <me@meorawr.io>